refactor: 컨트롤러가 응답 반환하도록 수정 및 helper 유틸 클래스 분리
